Schedule unit test's

+++ backend/src/modules/schedules/useCases/cancelShedule/cancelScheduleUseCase.js
@@ -6,11 +6,15 @@ class CancelScheduleUseCase {
     this.schedulesRepository = schedulesRepository;
   }
   async execute({ id_schedule, id_user }) {
-    const user = await this.prismaUsersRepository.findById({ id_user });
+    const user = await this.usersRepository.findById({ id_user });
     if (!user) {
       throw new AppError("User not found!!", 404);
     }
 
+    const schedule = await this.schedulesRepository.findById({ id_schedule });
+    if (!schedule) {
+      throw new AppError("Schedule not found!!", 404);
+    }
     await this.schedulesRepository.removeUserSchedule({ id_schedule });
 
     return;
